Weather: Add initial implementation.

+++ fr_yhe_api/src/main/java/com/friendly_machines/fr_yhe_api/watchprotocol/IWatchBinder.kt
@@ -8,7 +8,7 @@ import com.friendly_machines.fr_yhe_api.commondata.WatchWearingArm
 interface IWatchBinder : IBinder {
     fun setProfile(height: Int, weight: Int, sex: WatchProfileSex, age: Byte, arm: WatchWearingArm?)
     fun setWeather(
-        weatherType: Short, temp: Byte, maxTemp: Byte, minTemp: Byte, dummy: Byte/*0*/, month: Byte, dayOfMonth: Byte, dayOfWeekMondayBased: Byte, location: String
+        weatherType: Int, temp: Byte, maxTemp: Byte, minTemp: Byte, dummy: Byte/*0*/, month: Byte, dayOfMonth: Byte, dayOfWeekMondayBased: Byte, location: String
     )
 
     fun setMessage(type: com.friendly_machines.fr_yhe_api.commondata.MessageTypeMed, time: Int, title: String, content: String)
